feat(perfbudget): creat initialAssetsCost with hardcoded limit
