+++ lib/Compilation.js
@@ -3,6 +3,7 @@
 	Author Tobias Koppers @sokra
 */
 var async = require("async");
+var path = require("path");
 
 var Tapable = require("tapable");
 var EntryModuleNotFoundError = require("./EntryModuleNotFoundError");
@@ -1030,3 +1031,95 @@ Compilation.prototype.checkConstraints = function() {
 		chunk.checkConstraints();
 	}.bind(this));
 };
+
+Compilation.prototype.getAssetsByChunks = function() {
+	var webpackStatsJson = this.getStats().toJson();
+	var chunks = getFilteredChunks(webpackStatsJson);
+
+	function getFilteredChunks(webpackStatsJson) {
+		return webpackStatsJson.chunks.filter(function(chunk) {
+			var chunkName = chunk.names[0];
+			// This chunk doesn't have a name. This script can't handled it.
+			if (chunkName === undefined) {
+				return false;
+			}
+			// Skip if the chunk should be lazy loaded
+			if (!chunk.initial) {
+				return false;
+			}
+
+			// Add otherwise
+			return true;
+		});
+	};
+
+	var assets = {
+		// Will contain all js & css files by chunk
+		chunks: {},
+		// Will contain all js files
+		js: [],
+		// Will contain all css files
+		css: [],
+		// Will contain the html5 appcache manifest files if it exists
+		manifest: Object.keys(this.assets).filter(function(assetFile) {
+			return path.extname(assetFile) === '.appcache';
+		})[0]
+	};
+
+	// Append a hash for cache busting
+	for(var i = 0; i < chunks.length; i++) {
+		var chunk = chunks[i];
+
+		debugger;
+
+		var chunkName = chunk.names && chunk.names[0] || ""; 
+
+		assets.chunks[chunkName] = {};
+
+		// Prepend the public path to all chunk files
+		var chunkFiles = [].concat(chunk.files).map(function(chunkFile) {
+			return chunkFile;
+		});
+
+		// Webpack outputs an array for each chunk when using sourcemaps
+		// But we need only the entry file
+
+		var entry = chunkFiles[0];
+		assets.chunks[chunkName].size = chunk.size;
+		assets.chunks[chunkName].entry = entry;
+		assets.chunks[chunkName].hash = chunk.hash;
+		assets.js.push(entry);
+
+		// Gather all css files
+		var css = chunkFiles.filter(function(chunkFile) {
+			// Some chunks may contain content hash in their names, for ex. 'main.css?1e7cac4e4d8b52fd5ccd2541146ef03f'.
+			// We must proper handle such cases, so we use regexp testing here
+			return /.css($|\?)/.test(chunkFile);
+		});
+		assets.chunks[chunkName].css = css;
+		assets.css = assets.css.concat(css);
+	}
+
+	// Duplicate css assets can occur on occasion if more than one chunk
+	// requires the same css.
+	
+	assets.css = uniques(assets.css);
+
+	return assets;
+};
+
+// TODO# Find better module or make more reusable
+function uniques(array) {
+	var result = [],
+		val, ridx;
+	outer:
+		for(var i = 0, length = array.length; i < length; i++) {
+			val = array[i];
+			ridx = result.length;
+			while(ridx--) {
+				if(val === result[ridx]) continue outer;
+			}
+			result.push(val);
+		}
+	return result;
+}
