+++ lib/Stats.js
@@ -325,8 +325,6 @@ Stats.prototype.toJson = function toJson(options, forToString) {
 		});
 	}
 
-	debugger;
-
 	if(showPerformance && compilation.compiler.options.target === "web") { 
 		obj.performance = compilation.compiler.options.performance;
 	}
@@ -487,6 +485,7 @@ Stats.jsonToString = function jsonToString(obj, useColors) {
 		colors.bold(obj.publicPath);
 		newline();
 	}
+
 	if(obj.assets && obj.assets.length > 0) {
 		var t = [
 			["Asset", "Size", "Chunks", "", "Chunk Names"]
